Allow setting SPN parameter (new in php-krb5 1.1.3)
The SPN parameter allow specifying the entry to authenticate with in
the keytab file.
Setting it to null keeps the old behavior where it guesses the FQDN,
but this is problematic when running as vhost or behind a reverse
proxy such. When set to '0' it authenticates with any entry in the keytab
file.

It's still compatible with php-krb5 < 1.1.3 if spn is set to null
(default)